DEMOS-30: de-genericized table defs
